Move logic out of tree provider
